Move ownership of sequences from callee to caller
This makes the C API more complex but greatly simplifies thread
synchronization and all the complications that come with it.

The sequences are now stored in a context that is owned by the caller
instead of being stored in mutable static structures owned by the
callee.

The callee must allocated a context, then reference the context on each
call, and finally explicitly free the context when done.

This has additional benefits of:

* Removing dependecy on the lazy_static crate
* Removing the need for Send and Sync traits on the Sequence trait and
  on any types that implement the Sequence trait
* Removing overhead of obtaining a lock on mutable static structures