Fix require
* Fix source relative paths
* Fix idempotency of require

Previous to this commit, the tests for these features were incorrect or
insuffient.

For source relative paths, the test was insuffient because the path used
in test was also in the search path.  Fix test by adding another layer
of require with a path that is source relative but not in the search
path.

For idempotency of require, the test was incorrect.  The test expected
the wrong value for the Rv.  Fix by changing value from 1 to 2.

Fix source relative paths and idempotency of require by adding calls to
enter_require and leave_require in the grammar action for require.

NOTE: The current implementation of idempotency of require is not
zero-cost as far as memory.  Each succesive require for same path will
result in a new Item::Multiple with an empty Vec.  There is room for
improvement here.