Fix seed format due to API change in rand
Seed format was changed yet again.  This time from one extreme (u64) to
the other extreme (u8).  This extreme is even more painful that the
switch from u32 to u64.  This change in rand also makes use of `unsafe`
code to change representation of data in memory.  So what was wrong with
u32 again??

rand commit that broke rvs:
https://github.com/dhardy/rand/commit/888c5d2336d178bad0af6960a234ecd3af76dd3b