Better error handling on require error
Replace panics with bubble up error handling.

This is not done "the Rust way".  Due to abstration by rust-peg, it is
not possible to return errors via a Result return type.  Rust-peg
already uses its own Result return type.  It cannot be used for custom
errors.  To workaround this, errors are reported via insertion of
a RequireError into the AST.  The AST is then processed for presence of
RequireErrors during flattening.

Also flesh out library errors types for rvs-parser and rvs.