Fix evaluation after error
Previous to this change, if the REPL encountered a parse/transform
error, it would still attempt to evaluate the most recently parsed
variable.  But because of the error, no new variable would be added.
Error on first statement would result in a panic.  Error on second or
later statements would result in previous statement being re-evaluated.

Fix by moving parse and eval into a function and adding early return on
parse error.