Implment lazy evaluation of RHS variables
This implements lazy evaluation for both:

* Plain RHS variables
* Variables in function arguments

It does so by adding an intermediate representation (IR) between the AST
vector and the object model.  The IR is a map representation of the AST.
Variables are uniquified by name while keeping the order in which they
were first seen.  The IR is then transformed to the object model.
RHS variable lookups are performed during this transformation.  The
lookups find the most recent (aka last) definition of the variables.